Pressing enter while input buffer is active splits the text into two, like making a new line.
Currently allowing TextInput#add to take an optional parameter for 
where the user clicked, in order to set the position of the caret.
This parameter is currently OPTIONAL because it's awkward to figure
out a position to send, when you really want to just put the caret
at the beginning of the Text (as is the case when you press enter).